Moving TT tests to their own folder
In prep for adding in the Anaxym ones
