+++ .github/workflows/playwright regression.yml
@@ -1,4 +1,4 @@
-name: Playwright Tests
+name: Playwright Tests - Talent Ticker
 on:
   pull_request:
     branches: [ main ]
@@ -35,7 +35,7 @@ jobs:
         echo 'MAILOSAUR_PASSWORD="${{ secrets.MAILOSAUR_PASSWORD }}"' >> .env
         cat .env
     - name: Run Playwright tests
-      run: npx playwright test tests/general_regression/*
+      run: npx playwright test tests/talent_ticker/general_regression/*
     - uses: actions/upload-artifact@v2
       if: always()
       with:
