+++ playwright/tests/talent_ticker/email_regression/company_vacancy_contact_gmail.js
@@ -1,4 +1,4 @@
-const config = require('../../playwright.config.js');
+const config = require('../../../playwright.config.js');
 const { test, expect } = require('@playwright/test');
 const assert = require('chai').assert;
 const Mailosaur = require('mailosaur');
@@ -6,9 +6,9 @@ require('dotenv').config();
 
 test.use({ storageState: 'tests/state.json' });
 
-test("Contact Messaging From Sourcing Using Gmail Integration", async ({ page }) => {
+test("Contact Messaging From Vacancy Using Gmail Integration", async ({ page }) => {
   test.setTimeout(120000);
-
+  
   function Name_Alpha_Numeric() {
     var text = "";
     var possible = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789";
@@ -53,44 +53,65 @@ test("Contact Messaging From Sourcing Using Gmail Integration", async ({ page })
     page.click('button:has-text("Log In")')
   ]);
 
-  await expect(page).toHaveURL(config.use.baseURL + 'en-US/home');
+  await expect(page).toHaveURL(config.use.baseURL + 'home');
 
-  // Nav to Talent Sourcing
-  await page.click('[data-test="talentSourcingNavButton"]');
-  await expect(page).toHaveURL(config.use.baseURL + 'en-US/sourcing');
-  
-  // Fill [placeholder="e\.g\.\ Digital\ Designer"]
-  await page.fill('[placeholder="e\\.g\\.\\ Digital\\ Designer"]', 'test account');
-  // Click text=test account
-  await page.click('text=test account');
-  // Fill [placeholder="e\.g\.\ Monzo\ Bank"]
-  await page.fill('[placeholder="e\\.g\\.\\ Monzo\\ Bank"]', 'selligence');
-  // Click li[role="option"]:has-text("Selligence")
-  await page.click('li[role="option"]:has-text("Selligence")');
-  // Click [data-testid="search-btn"]
-  await page.click('[data-testid="search-btn"]');
+  // Click [data-test="vacanciesNavButton"]
+  await page.click('[data-test="vacanciesNavButton"]');
+  await expect(page).toHaveURL('https://staging.talentticker.ai/vacancies');
+
+   // Click [placeholder="Search"]
+   await page.click('[placeholder="Search"]');
+   // Click div[role="listbox"] >> text=Saved
+   await page.click('div[role="listbox"] >> text=Saved');
+   // Click [placeholder="Search"]
+   await page.click('[placeholder="Search"]');
+   // Click [data-test="pageTemplate"] >> text=Selligence
+   await page.click('[data-test="pageTemplate"] >> text=Selligence');
+   await expect(page).toHaveURL('https://staging.talentticker.ai/news');
+   // Click [data-test="vacanciesTabButton"] >> text=Vacancies
+   await page.click('[data-test="vacanciesTabButton"] >> text=Vacancies');
+   await expect(page).toHaveURL('https://staging.talentticker.ai/vacancies');
+
+  expect(await page.innerText('strong')).toContain("Selligence");
+
+  // Click [data-testid="placed-event-title"]
+  await page.locator(':nth-match(h5[data-testid="placed-event-title"], 1)').click();
+
+  // Click [data-testid="contacts-section"] >> text=Clear all
+  await page.click('[data-testid="contacts-section"] >> text=Clear all');
+
+  // Fill [placeholder="Search\ by\ name\,\ role\ or\ location"]
+  await page.fill('[placeholder="Search\\ by\\ name\\,\\ role\\ or\\ location"]', 'Test Account Manager');
+  // Click text=“Tom Jones”
+  await page.click('text=“Test Account Manager”');
   // Check input[type="checkbox"]
   await page.check('input[type="checkbox"]');
   // Click text=emailMessage (1)
   await page.click('text=emailMessage (1)');
+
   const randomName = Name_Alpha_Numeric();
   // Fill [aria-label="subject"]
   await page.fill('[aria-label="subject"]', randomName);
   // Click text=Send Message
   await page.click('text=Send Message');
+
+  await expect(
+    page.locator('[data-testid="message-send-success"]')
+  ).toContainText("Message successfully sent");
+
   // Click button:has-text("close")
-  await page.click('button:has-text("close")');
+  await page.click('[data-testid="close-drawer-button"]');
 
-  await page.click('[data-test="pageTemplate"] >> text=Clear All');
-
+  await page.click('[aria-label="Close"]');
+  
   // Click [data-test="contatcsNavButton"]
   await page.click('[data-test="contatcsNavButton"]');
-  await expect(page).toHaveURL(config.use.baseURL + 'en-US/outbox');
+  await expect(page).toHaveURL('https://staging.talentticker.ai/outbox');
 
   expect(await page.innerText('h1')).toContain("Outbox");
 
   await page.click('[data-test="contatcsNavButton"]');
-  await expect(page).toHaveURL(config.use.baseURL + 'en-US/outbox');
+  await expect(page).toHaveURL('https://staging.talentticker.ai/outbox');
 
   expect(await page.innerText('h1')).toContain("Outbox");
 
@@ -103,7 +124,7 @@ test("Contact Messaging From Sourcing Using Gmail Integration", async ({ page })
           mailFound = true;
       } catch (e) {
           i++;
-          await page.click('[id="reload-outbox"]');
+          await page.click('[data-testid="reload-list-btn"]');
           if (i == 10){
             throw e
           }
